+++ src/watchdog/observers/read_directory_changes.py
@@ -64,7 +64,21 @@ class WindowsApiEmitter(EventEmitter):
     def __init__(self, event_queue, watch, timeout=DEFAULT_EMITTER_TIMEOUT):
         EventEmitter.__init__(self, event_queue, watch, timeout)
         self._lock = threading.Lock()
-        self._handle = get_directory_handle(watch.path)
+        self._handle = None
+
+    def run(self):
+        """
+        Start the looking for changes.
+        """
+        try:
+            self._handle = get_directory_handle(self._watch.path)
+        except Exception as error:
+            self._start_error = error
+            return
+        finally:
+            self.ready.set()
+
+        return EventEmitter.run(self)
 
     def on_thread_stop(self):
         close_directory_handle(self._handle)
@@ -75,7 +89,7 @@ class WindowsApiEmitter(EventEmitter):
             last_renamed_src_path = ""
             for winapi_event in winapi_events:
                 src_path = os.path.join(self.watch.path, winapi_event.src_path)
-                
+
                 if winapi_event.is_renamed_old:
                     last_renamed_src_path = src_path
                 elif winapi_event.is_renamed_new: