+++ src/watchdog/tests/observers/test_inotify.py
@@ -1,8 +1,11 @@
 """
 Test for inotify emitter and observer.
 """
+from Queue import Queue
+import threading
 
 from watchdog.tests import WatchdogTestCase
+from watchdog.tests.observers.emitter_mixin import EmitterSystemMixin
 
 # Inotify is only supported on Linux.
 if WatchdogTestCase.os_name != 'linux':
@@ -12,6 +15,8 @@ if WatchdogTestCase.os_name != 'linux':
 if 'rhel-4' in WatchdogTestCase.getHostname():
     raise WatchdogTestCase.skipTest()
 
+# Inotify is imported later.
+from watchdog.observers.inotify import InotifyEmitter
 from watchdog.observers.inotify_buffer import STOP_EVENT
 
 
@@ -24,9 +29,11 @@ class InMemoryInotifyBuffer(object):
             queue = []
         self.queue = queue
         self._inotify = None
+        self.ready = threading.Event()
 
     def start(self):
         self._inotify = object()
+        self.ready.set()
 
     def close(self):
         self._inotify = None
@@ -36,28 +43,52 @@ class InMemoryInotifyBuffer(object):
         return self.queue.pop()
 
 
-class InotifyEmitterTestCase(WatchdogTestCase):
+class TestInotifyEmitter(WatchdogTestCase, EmitterSystemMixin):
     """
     Unit tests for InotifyEmitter.
     """
 
     def setUp(self):
-        super(InotifyEmitterTestCase, self).setUp()
-        from watchdog.observers.inotify import InotifyEmitter
+        super(TestInotifyEmitter, self).setUp()
 
-        self.emitter_queue = []
-        self.sut = InotifyEmitter(
+        self.emitter_queue = Queue()
+        # Configure emitter for temp folder.
+        self.sut = self.makeEmitter()
+        self.buffer = InMemoryInotifyBuffer()
+
+    def makeEmitter(self, path=None):
+        """
+        Return a new emitter monitoring `path`.
+        """
+        return InotifyEmitter(
             event_queue=self.emitter_queue,
-            watch=self.ObservedWatch(),
+            watch=self.ObservedWatch(path=path),
             timeout=0,
             )
-        self.buffer = InMemoryInotifyBuffer()
 
     def patchBuffer(self):
         """
         Patch the emitter with a fake buffer.
         """
         self.sut._inotify = self.buffer
+
+    def test_start_failure(self):
+        """
+        Emitter will be ready with last error set when failing to start.
+
+        This is an unit test,
+        """
+        error = AssertionError('fail-to-start')
+        self.patchBuffer()
+        self.buffer.start = self.Mock(side_effect=[error])
+
+        self.sut.start()
+
+        # Low level buffer is not ready but high level emitter is ready
+        # due to start error.
+        self.assertIsFalse(self.buffer.ready.is_set())
+        self.assertIs(error, self.sut.start_error)
+        self.assertIsTrue(self.sut.ready.is_set())
 
     def test_queue_events_stop(self):
         """
@@ -69,4 +100,4 @@ class InotifyEmitterTestCase(WatchdogTestCase):
 
         self.sut.queue_events()
 
-        self.assertIsEmpty(self.emitter_queue)
+        self.assertTrue(self.emitter_queue.empty())